+++ src/LightVx/Validators/LengthValidator.cs
@@ -1,7 +1,9 @@
+using System;
+
 namespace LightVx.Validators
 {
     /// <summary>
-    ///     Validate text with length options
+    ///     Validate text or arrays with length options
     /// </summary>
     public class LengthValidator : ValidatorBase
     {
@@ -27,6 +29,12 @@ namespace LightVx.Validators
 
         protected override void Validate()
         {
+            if (Input != null && Input is Array)
+            {
+                ValidateArray();
+                return;
+            }
+
             if (_Input == null && _minOccurance == 0)
             {
                 Succeed();
@@ -62,6 +70,29 @@ namespace LightVx.Validators
                             _maxOccurance);
                 Fail(message);
             }
+        }
+
+        private void ValidateArray()
+        {
+            if(((Array)Input).Length < _minOccurance)
+            {
+                Fail(
+                    string.Format("is not valid. Must have a length greater than {0}.",
+                        _minOccurance));
+                return;
+            }
+
+            if (_maxOccurance.HasValue)
+            {
+                if (_maxOccurance.Value > ((Array)Input).Length)
+                {
+                    Fail(
+                        string.Format("is not valid. Must have a length of between {0} and {1}.",
+                            _minOccurance, _maxOccurance));
+                    return;
+                }
+            }
+            Succeed();
         }
     }
 }