+++ src/Tests/LightVx.Tests/ObjectValidator/Customer.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LightVx.Tests.ObjectValidator
+{
+    /// <summary>
+    /// A sample class for Object Validation
+    /// </summary>
+    public class Customer
+    {
+        public string Name { get; set; }
+        public DateTime DOB { get; set; }
+    }
+}